Move files to lib/
